requirements.txt
python-dotenv >= 0.7.1
numpy >= 1.13.3
tensorflow >= 0.8.0
